Add Weapon Audio || Add Player Move Audio || Add Audio to Interactables
This Commit adds weapon audio to each weapon and adds audio for player movement and interactables.

Due to the fire rate of weapons like the machine gun, a pool of audio sources needed to be created. That way, multiple shoot sounds can overlap with each other without overriding the previous sound.